[Resize content][1/n] Break horizontal scrolling area into final-ish hierarchy
Summary
---

This PR begins a PR stack that implements the resize of the lanes area.

In this PR, the view hierarchy is overhauled to more closely resemble
the final one. This is the new view hierarchy:

- View (vertically stacked layout) (root view)
  - HorizontalPanAndZoomView
    - View (vertically stacked layout)
      - TimeAxisMarkersView
      - ReactEventsView
  - View (vertically stacked layout) (to be replaced with
    ResizableVerticalSplitView)
    - HorizontalPanAndZoomView
      - VerticalScrollView
        - ReactMeasuresView
    - HorizontalPanAndZoomView
      - VerticalScrollView
        - FlamechartView

This PR must be landed with the PRs stacked above it in order not to
break master.

Test Plan
---

* `yarn lint`
* `yarn flow`: no errors in changed code
* `yarn test`

Known issues
---

* There are now 3 independent horizontal pan/zoom areas. This is
  intentional; it'll be fixed in a future PR in this stack.
* Flamechart no longer scrolls vertically. This will be fixed in a
  future PR in this stack.